requirements.txt
Flask==3.0.0
Authlib==1.2.1
requests==2.31.0
mysql-connector-python==8.2.0
